pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-unfold>=0.66.0",
    "djangorestframework>=3.16.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
